Refactoring: Separación de código para recuperar imágenes hacia una nueva clase.
