+++ app/src/main/res/layout/fragment_account.xml
@@ -233,6 +233,13 @@
             />
 
     </LinearLayout>
+
+    <androidx.fragment.app.FragmentContainerView
+        android:id="@+id/fragmentContainer"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_weight="1"/>
+
 </androidx.constraintlayout.widget.ConstraintLayout>
 
 </ScrollView>
